refactor: move cppad_py -> test_ad and adapt to empty __init__ for modules.
