+++ test_ad/test_gaussian_pdf.py
@@ -1,9 +1,11 @@
 import numpy
 import cppad_py
-import curvefit
-import a_functions
+import curvefit.core.utils
+import curvefit.core.effects2params
+import curvefit.core.objective_fun
+from test_ad import a_functions
 #
-def test_dgaussian_pdf() :
+def test_gaussian_pdf() :
     eps99  = 99.0 * numpy.finfo(float).eps
     #
     # test values for t, param
@@ -19,17 +21,17 @@ def test_dgaussian_pdf() :
         for j in range( param.shape[1] ) :
             aparam[i][j] = cppad_py.a_double( param[i][j] )
     # -----------------------------------------------------------------------
-    # f(t) = gaussian_pdf(t, param)
+    # f(t) = gaussian_cdf(t, param)
     at = cppad_py.independent(t)
-    ay = a_functions.a_gaussian_pdf(at, aparam)
+    ay = a_functions.a_gaussian_cdf(at, aparam)
     f  = cppad_py.d_fun(at, ay)
     #
-    # g(t) = dgaussian_pdf(t, param)
+    # g(t) = d/dt gaussian_cdf(t, param)
     at = cppad_py.independent(t)
-    ay = a_functions.a_dgaussian_pdf(at, aparam)
+    ay = a_functions.a_gaussian_pdf(at, aparam)
     g  = cppad_py.d_fun(at, ay)
     #
-    # check a_dgaussian_pdf
+    # check a_gaussian_pdf
     f.forward(0, t)
     g0  = g.forward(0, t)
     dt  = a_functions.constant_array((t.size,), 0.0)
@@ -39,3 +41,11 @@ def test_dgaussian_pdf() :
         rel_error = g0[i] / df[i] - 1.0
         dt[i]     = 0.0
         assert abs(rel_error) < eps99
+    # -----------------------------------------------------------------------
+    # check a_ln_dgaussain_cdf
+    at = cppad_py.independent(t)
+    ay = a_functions.a_ln_gaussian_pdf(at, aparam)
+    f  = cppad_py.d_fun(at, ay)
+    f0 = f.forward(0, t)
+    rel_error = f0 / numpy.log(g0) - 1
+    assert all( abs(rel_error) < eps99 )
